requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements-dev.in
#
aiohttp==3.8.3
    # via black
aiosignal==1.2.0
    # via aiohttp
astroid==2.12.11
    # via pylint
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via
    #   aiohttp
    #   pytest
black[d]==22.10.0
    # via -r requirements-dev.in
charset-normalizer==2.1.1
    # via aiohttp
click==8.1.3
    # via black
colorama==0.4.5
    # via
    #   click
    #   pylint
    #   pytest
    #   tox
coverage[toml]==6.5.0
    # via pytest-cov
dill==0.3.5.1
    # via pylint
distlib==0.3.6
    # via virtualenv
filelock==3.8.0
    # via
    #   tox
    #   virtualenv
flake8==5.0.4
    # via -r requirements-dev.in
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via yarl
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via pylint
lazy-object-proxy==1.7.1
    # via astroid
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
mypy==0.982
    # via -r requirements-dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.3
    # via
    #   pytest
    #   tox
pathspec==0.10.1
    # via black
platformdirs==2.5.2
    # via
    #   black
    #   pylint
    #   virtualenv
pluggy==1.0.0
    # via
    #   pytest
    #   tox
py==1.11.0
    # via
    #   pytest
    #   tox
pycodestyle==2.9.1
    # via flake8
pyflakes==2.5.0
    # via flake8
pylint==2.15.4
    # via
    #   -r requirements-dev.in
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.5.3
    # via -r requirements-dev.in
pylint-plugin-utils==0.7
    # via pylint-django
pyparsing==3.0.9
    # via packaging
pytest==7.1.3
    # via
    #   -r requirements-dev.in
    #   pytest-cov
pytest-cov==4.0.0
    # via -r requirements-dev.in
six==1.16.0
    # via tox
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   mypy
    #   pylint
    #   pytest
    #   tox
tomlkit==0.11.5
    # via pylint
tox==3.26.0
    # via -r requirements-dev.in
typing-extensions==4.4.0
    # via mypy
virtualenv==20.16.5
    # via tox
wrapt==1.14.1
    # via astroid
yarl==1.8.1
    # via aiohttp
